CLion xmls generation added Few linter warnings fixed Disabled del command output by default Removed old test file as it no longer used Added new test file as private file Cleanup of .gitignore and libraries which don't have examples

+++ clion.py
@@ -7,20 +7,24 @@ from . import conan
 from . import log
 from . import commons
 
-def _log_error_and_throw(mgs: str) -> typing.NoReturn:
-    log.error(mgs)
-    raise Exception(mgs)
+def _error_log_and_die(msg: str) -> typing.NoReturn:
+    log.error(msg)
+    exit(-1)
 
-# <component name="CMakeSettings">
-#     <configurations>
-#       <configuration PROFILE_NAME="bt.Debug" ENABLED="true" GENERATION_DIR=".workspace/clion/debug" CONFIG_NAME="Debug" TOOLCHAIN_NAME="Visual Studio" GENERATION_OPTIONS="-DCMAKE_PREFIX_PATH=&quot;I SEE THAT&quot;" BUILD_OPTIONS="-j16" NO_GENERATOR="true" />
-#       <configuration PROFILE_NAME="bt.Release" ENABLED="true" GENERATION_DIR=".workspace/clion/release" CONFIG_NAME="Release" TOOLCHAIN_NAME="Visual Studio" GENERATION_OPTIONS="-DCMAKE_PREFIX_PATH=&quot;I SEE THAT&quot;" BUILD_OPTIONS="-j16" NO_GENERATOR="true" />
-#       <configuration PROFILE_NAME="conan-default" ENABLED="false" FROM_PRESET="true" GENERATION_DIR="$PROJECT_DIR$/.workspace/conan2_home/dependencies/debug/build" />
-#       <configuration PROFILE_NAME="conan-default - conan-debug" ENABLED="false" FROM_PRESET="true" GENERATION_DIR="$PROJECT_DIR$/.workspace/conan2_home/dependencies/debug/build" />
-#     </configurations>
-#   </component>
 
 def create_build_tools_configurations(
+    workspace_dir: str,
+    project_dir: str,
+    ninja_exe: str | None = None,
+    vcpkg_dependencies: list[str] | None = None
+):
+    log.info("Creating CLion Project XML File with CMake settings...")
+    _create_idea_project_xml(workspace_dir, project_dir, ninja_exe, vcpkg_dependencies)
+    log.info("Creating CLion basic dictionary...")
+    _create_idea_dictionary_xml(project_dir)
+    log.info("Creating CLion files finished successfully!")
+
+def _create_idea_project_xml(
     workspace_dir: str,
     project_dir: str,
     ninja_exe: str | None = None,
@@ -40,11 +44,11 @@ def create_build_tools_configurations(
             """).strip())
 
     tree = xml.etree.ElementTree.parse(IDEA_WORKSPACE_FILE)
-    project = tree.find('.')
-    if project == None:
-        _log_error_and_throw('<project> tag not found in .idea/workspace.xml. Process of creating config aborted.')
+    project = tree.getroot()
+    if project.tag != "project":
+        _error_log_and_die('<project> tag not found in .idea/workspace.xml. Process of creating config aborted.')
     if project.attrib['version'] != "4":
-        _log_error_and_throw("<project> version is different from 4, which is not supported")
+        _error_log_and_die("<project> version is different from 4, which is not supported")
 
 
     cmake_settings = project.find('./component[@name="CMakeSettings"]')
@@ -92,8 +96,6 @@ def create_build_tools_configurations(
                 "-j16"
             ]
 
-        #set_target_properties(${_LIB_NAME} PROPERTIES IMPORTED_LOCATION ${CONAN_FOUND_LIBRARY} IMPORTED_NO_SONAME ${no_soname_mode})
-
         new_config: dict[str, str] = {
             "PROFILE_NAME" : mode,
             "ENABLED" : "true",
@@ -106,7 +108,9 @@ def create_build_tools_configurations(
         }
 
         xml.etree.ElementTree.SubElement(configurations_element, 'configuration', new_config)
+
     tree.write(IDEA_WORKSPACE_FILE)
+    
     log.warn("----------------- [CLION IMPORTANT NOTE] ---------------------")
     log.warn("There is no way to enforce CMake version inside CLion single project")
     log.warn("If you want to use build_tools downloaded CMake version you have to")
@@ -115,20 +119,37 @@ def create_build_tools_configurations(
     log.warn("You have to setup: Preferences > Build, Execution, Deployment > Toolchains > YOUR_TOOLCHAIN > CMake")
     log.warn("----------------- [CLION IMPORTANT NOTE] ---------------------")
 
-
+def _create_idea_dictionary_xml(project_dir: str):
     IDEA_DICTIONARY_FILE = f'{project_dir}/.idea/dictionaries/project.xml'
     os.makedirs(os.path.dirname(IDEA_DICTIONARY_FILE), exist_ok=True)
-    with open(IDEA_DICTIONARY_FILE, "w", encoding="UTF-8") as file:
-        file.write(textwrap.dedent("""
-            <component name="ProjectDictionaryState">
-                <dictionary name="project">
-                    <words>
-                        <w>ctre</w>
-                    </words>
-                </dictionary>
-            </component>
-        """))
+    if not os.path.exists(IDEA_DICTIONARY_FILE):
+        with open(IDEA_DICTIONARY_FILE, "w", encoding="UTF-8") as file:
+            file.write(textwrap.dedent("""
+                <component name="ProjectDictionaryState">
+                    <dictionary name="project">
+                        <words>
+                        </words>
+                    </dictionary>
+                </component>
+            """).strip())
 
-    
 
+    tree = xml.etree.ElementTree.parse(IDEA_DICTIONARY_FILE)
+    project_dictionary_state = tree.getroot()
+    if project_dictionary_state.tag != "component" and project_dictionary_state.attrib['name'] != "ProjectDictionaryState":
+        log.warn('<component name="ProjectDictionaryState"> Not found in dictionary file. Omitting dictionary generation')
+        return
     
+    project_dictionary = project_dictionary_state.find('./dictionary[@name="project"]')
+    if project_dictionary == None:
+        project_dictionary = xml.etree.ElementTree.SubElement(project_dictionary_state, 'dictionary', attrib={'name' : "project"})
+    
+    project_dictionary_words = project_dictionary.find('./words')
+    if project_dictionary_words == None:
+        project_dictionary_words = xml.etree.ElementTree.SubElement(project_dictionary, 'words')
+
+    for word_text in ["ctre"]:
+        word = xml.etree.ElementTree.SubElement(project_dictionary_words, 'word')
+        word.text = word_text
+
+    tree.write(IDEA_DICTIONARY_FILE)
